chore: rename root reducer
